# Conflicts:
#	app/src/main/java/com/multimoney/multimoney/presentation/ui/crypto/market/MarketScreen.kt
#	app/src/main/java/com/multimoney/multimoney/presentation/uielement/CustomInformativeText.kt

+++ domain/src/main/java/com/multimoney/domain/repository/CryptoRepository.kt
@@ -74,7 +74,7 @@ interface CryptoRepository {
         market: String,
         commissionAmount: Double,
         taxAmount: Double,
-        accountToken: Double,
+        accountToken: Long,
         exchangeRate: Double,
         idBrand: Int,
         user: String,
